more fixes

+++ srcs/ft_pipex.c
@@ -6,7 +6,7 @@
 /*   By: aarrien- <aarrien-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/17 11:59:42 by aarrien-          #+#    #+#             */
-/*   Updated: 2023/02/16 17:08:48 by aarrien-         ###   ########.fr       */
+/*   Updated: 2023/02/20 12:21:03 by aarrien-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,11 +20,13 @@ void	ft_child(t_pipex *pipex, int i, t_env **list)
 	if (i != pipex->npipes)
 		dup2(pipex->fds[i][1], STDOUT_FILENO);
 	ft_close(pipex, pipex->npipes);
-	if (pipex->cmds[i])
-		execve(pipex->cmds[i][0], pipex->cmds[i], pipex->env);
-	ft_error_msg("pipex: command not found: ", pipex->cmds[i][0]);
-	unlink(".temp.txt");
-	exit(0);
+	if (choose_command_child(pipex->cmds[i], list) == 0)
+	{
+		if (pipex->cmds[i])
+			execve(pipex->cmds[i][0], pipex->cmds[i], pipex->env);
+		ft_error_msg("pipex: command not found: ", pipex->cmds[i][0]);
+		unlink(".temp.txt");
+	}
 	exit(0);
 }
 
@@ -40,18 +42,19 @@ int	ft_fork(t_pipex *pipex, t_env **list)
 	temp2 = dup(STDOUT_FILENO);
 	dup2(pipex->inout_fd[0], STDIN_FILENO);
 	dup2(pipex->inout_fd[1], STDOUT_FILENO);
+	if (choose_command_father(pipex->cmds[i], list) != 0)
+		i++;
 	while (pipex->npipes - i + 1 > 0)
 	{
-		if (choose_command(pipex->cmds[i], list) == 0)
-		{
-			pid = fork();
-			if (pid == 0)
-				ft_child(pipex, i, list);
-			ft_close(pipex, i);
-			waitpid(pid, NULL, 0);
-		}
+		signal(SIGINT, exec_handle_signal);
+		pid = fork();
+		if (pid == 0)
+			ft_child(pipex, i, list);
+		ft_close(pipex, i);
+		waitpid(pid, NULL, 0);
 		i++;
 	}
+	signal(SIGINT, handle_signal);
 	dup2(temp1, STDIN_FILENO);
 	dup2(temp2, STDOUT_FILENO);
 	return (0);
@@ -103,6 +106,7 @@ void	fill_cmds(t_input	*input)
 	cmd_pos = 0;
 	input->pipex->cmds = malloc((input->pipex->npipes + 2) * sizeof(char **));
 	input->pipex->cmds[input->pipex->npipes + 1] = NULL;
+	input->pipex->cmds[0] = NULL;
 	while (input->commd[i])
 	{
 		if (*input->commd[i] == '|' && i++)
@@ -120,14 +124,16 @@ void	fill_cmds(t_input	*input)
 
 // joins the command with its absolute path
 // if he doesn't find it does nothing
-char	*attach_path(char *cmd)
+char	*attach_path(char *cmd, t_env **list)
 {
 	char	*p_cmd;
 	char	*aux;
 	char	**splited;
 	int		i;
 
-	p_cmd = getenv("PATH");
+	p_cmd = find_var("PATH", list);
+	if (!p_cmd)
+		return (cmd);
 	splited = ft_split(p_cmd, ':');
 	i = 0;
 	while (splited[i])
@@ -169,9 +175,11 @@ void	ft_open(t_input	*input)
 	if (input->otxt[1])
 	{
 		if (ft_strlen(input->otxt[0]) == 2)
-			input->pipex->inout_fd[1] = open(input->otxt[1], O_CREAT | O_WRONLY | O_APPEND, 0644);
+			input->pipex->inout_fd[1] = open(input->otxt[1],
+					O_CREAT | O_WRONLY | O_APPEND, 0644);
 		else
-			input->pipex->inout_fd[1] = open(input->otxt[1], O_CREAT | O_WRONLY | O_TRUNC, 0644);
+			input->pipex->inout_fd[1] = open(input->otxt[1],
+					O_CREAT | O_WRONLY | O_TRUNC, 0644);
 	}
 	else
 		input->pipex->inout_fd[1] = 1;
@@ -189,13 +197,11 @@ int	ft_pipex(t_input *input, t_env **list)
 	j = 0;
 	while (input->pipex->cmds[j])
 	{
-		i = 0;
-		while (input->pipex->cmds[j][i])
-		{
+		i = -1;
+		while (input->pipex->cmds[j][++i])
 			if (i == 0 && input->pipex->cmds[j][i][0] != '/')
-				input->pipex->cmds[j][i] = attach_path(input->pipex->cmds[j][i]);
-			i++;
-		}
+				input->pipex->cmds[j][i] = \
+					attach_path(input->pipex->cmds[j][i], list);
 		j++;
 	}
 	i = 0;
@@ -203,8 +209,7 @@ int	ft_pipex(t_input *input, t_env **list)
 	while (input->pipex->npipes - i > 0)
 	{
 		input->pipex->fds[i] = malloc(2 * sizeof(int));
-		pipe(input->pipex->fds[i]);
-		i++;
+		pipe(input->pipex->fds[i++]);
 	}
 	ft_open(input);
 	ft_fork(input->pipex, list);
